hamal-techtrend/requirements.txt
# Core LLM/Agent Libraries
anthropic==0.18.1
openai
langchain==0.1.9
langgraph==0.0.26
langchain-anthropic==0.1.4
langchain-openai

# Web Search
tavily-python

# Utilities
python-dotenv==1.0.1
click==8.1.7
pydantic==2.6.1

# Optional (for embeddings/similarity search)
sentence-transformers==2.3.1
numpy==1.26.3

# Optional (for PDF parsing if uploading charters)
pypdf2==3.0.1
litellm==1.59.8

# Backend / Deployment
fastapi==0.109.2
uvicorn==0.27.1
